chore: rename sourcesets
