+++ gradle.properties
@@ -11,18 +11,21 @@ java_version = 17
 # Kotlin settings
 kotlin.code.style = official
 kotlin.incremental = true
-kotlin_version = 1.8.21
+kotlin_version = 1.8.22
+
+# Fabric settings
+fabric.loom.multiProjectOptimisation = true
 
 # Dependency versions
-fabric_language_kotlin_version = 1.9.4+kotlin.1.8.21
-fabric_loader_version = 0.14.19
+fabric_language_kotlin_version = 1.9.5+kotlin.1.8.22
+fabric_loader_version = 0.14.21
 fabric_loom_version = 1.2-SNAPSHOT
-fabric_version = 0.79.0+1.19.4
-minecraft_version = 1.19.4
-run_paper_version = 2.0.1
+fabric_version = 0.83.1+1.20.1
+minecraft_version = 1.20.1
+run_paper_version = 2.1.0
 shadow_version = 8.1.1
-paper_api_version = 1.19.4-R0.1-SNAPSHOT
-paperweight_version = 1.5.4
-plugin_yml_version = 0.5.3
+paper_api_version = 1.20-R0.1-SNAPSHOT
+paperweight_version = 1.5.5
+plugin_yml_version = 0.6.0
 velocity_api_version = 3.1.1
-yarn_mappings_version = 1.19.4+build.2
+yarn_mappings_version = 1.20.1+build.3
