+++ projects/example-fabric/src/server/resources/assets/fabric_example/icon.png
