+++ projects/example-fabric/src/server/resources/fabric.mod.json
