+++ projects/example-fabric/src/server/resources/fabric_example.mixins.json
@@ -1,11 +1,11 @@
 {
   "required": true,
-  "package": "io.layux.fabricexample.mixin",
+  "package": "io.layux.example.mixin",
   "compatibilityLevel": "JAVA_17",
   "injectors": {
     "defaultRequire": 1
   },
   "client": [
-    "ExampleMixin"
+    "io.layux.example.mixin.ExampleMixin"
   ]
 }
