+++ projects/example-paper/build.gradle
@@ -26,7 +26,7 @@ bukkit {
     description = project.description
     load = BukkitPluginDescription.PluginLoadOrder.STARTUP
     main = "${group}.ExamplePlugin"
-    apiVersion = "1.19"
+    apiVersion = "1.20"
     authors = ["Layux"]
 }
